jira_integration_app/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jira==3.8.0
jiter==0.8.2
oauthlib==3.2.2
openai==1.63.2
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
